refactor(agents): Break up client side prompt tool handlers
